Function renaming.
ICU #defines all the function names to add version suffixes as a crude way
to allow multiple versions to coexist. We don't want these suffixes in the
Rust names. Bindgen itself does not support this technique, but since the
bindings are very regular, we can easily replace the names with a regex.