Add default icu4c installation dir to PKG_CONFIG_PATH on Darwin
brew by default installs icu4c to /usr/local/opt/icu4c, so add it to the
end of PKG_CONFIG_PATH so it gets found without special action on the part
of the user.

Note that icu4c is already brewed on the Travis Darwin builders, so the
brew is not actually needed.